syscall: add support for getting uid/gid/pid and related numbers
This is for compatibility with the existing syscall package and in fact
these functions are implemented in the baremetal syscall version. The
numbers will most likely be fake for WebAssembly but that's the job of
wasi-libc to sort out.

I'm not sure how to add tests for these, as they are normally unknown by
definition (unlike command line arguments and environment variables
which can be set by a test program).